refactor: win impl
